- removed unused method for PCA9632 - MeasureTypes is now uint16_t in all constructors - added lux and tof sensors - added all i2c addresses as defines to according sensor implementation - updated platformio.ini with autoload lib functionality

+++ src/SensorBME680.cpp
@@ -5,12 +5,12 @@
 #define STATE_SAVE_PERIOD  UINT32_C(360 * 60 * 1000) // 360 minutes - 4 times a day
 #define EEPROM_BME680_START_ADDRESS 0xC80
 
-SensorBME680::SensorBME680(uint8_t iMeasureTypes, uint8_t iAddress, bme680_delay_fptr_t iDelayCallback)
+SensorBME680::SensorBME680(uint16_t iMeasureTypes, uint8_t iAddress, bme680_delay_fptr_t iDelayCallback)
     : Sensor(iMeasureTypes, iAddress), Bsec(), mDelayCallback(iDelayCallback) {
     mEEPROM = new EepromManager(100, 5, sMagicWord);
 }
 
-SensorBME680::SensorBME680(uint8_t iMeasureTypes, uint8_t iAddress, bme680_delay_fptr_t iDelayCallback, uint8_t iMagicKeyOffset)
+SensorBME680::SensorBME680(uint16_t iMeasureTypes, uint8_t iAddress, bme680_delay_fptr_t iDelayCallback, uint8_t iMagicKeyOffset)
     : Sensor(iMeasureTypes, iAddress), Bsec(), mDelayCallback(iDelayCallback) {
     mEEPROM = new EepromManager(100, 5, sMagicWord);
     sMagicWord[0] ^= iMagicKeyOffset;
